storage-node: docs and minor refactoring of external promise util
